cleanup casting of conversions/adapters * rounding errors because (float)casting/conversion inbetween

+++ src/com/fr3ts0n/ecu/IntConversion.java
@@ -31,14 +31,13 @@ public class IntConversion
 	@Override
 	public Number memToPhys(long value)
 	{
-		return (float) value;
-
+		return value;
 	}
 
 	@Override
 	public Number physToMem(Number value)
 	{
-		return value.longValue();
+		return value;
 	}
 
 	@Override
